Ron comment: merge test suites, remove derive ready
